MB-15165: Download data files
Instead of placing the data and query files for the spatial views tests
manually somewhere, download them from an URL and store them on disk.
The current root directory of the downloaded files is `/tmp/spatial`.

Once a file is downloaded its SHA1 sum will be checked, to make sure the
download was successful. The files won't be re-downloaded if they already
exist.

Change-Id: I0c67ae6db0c0020b0a3df5a9c09f43610220f803