Add section information to settings
All settings now contain a `_section` property which has the section name
as value. This makes it possible to determine e.g. whether you are in the
access or the load phase.

From now on properties with a leading underscore are ones that are not
specified in the configuration file, but are generated in some other ways.

Tests that are run on nodes which have `n1ql` need to have a [n1ql] section
defined (leave it empty for default values).
Tests that are run on nodes which have `index` need to have a [secondary] section
defined (leave it empty for default values).

All tests that are run on the performance CI got checked if they need the
sections mentioned above and updated accordingly.

Change-Id: I66ad1b4309f892bcc23adff88b0474aa31b77e7a