Don't populate sections that are not defined
If a section is not defined in the test configuration it won't be accessible from
within the Python code. If a section should be populated with the default values
it needs to be explicitly added in the configuration file. An example would be
the `[bucket]` section which is part of all test files, but often without any
properties set. It will then use the hard-coded defaults.

Currently every test file has these mandatory sections:

 - [bucket]
 - [cluster]
 - [stats]
 - [test_case]
 - [worker_settings]

If one of them isn't defined the test will interrupt with an error.

Change-Id: I99f6a91fd3a3b4e9332aaadc610b4dbbf68615ce