MB-15514: Port mapreduce views simple latency tests
The mapreduce views latency tests now use the new classes. This makes
the configuration as well as the internal names a bit easier to understand.

The behaviour of the `stale=false` tests changes a bit. Before this port the
tests were running the background load with `stale=update_after` and only the
requests for the measurements with `stale=false`. The ported tests runs all
requests, the load as well as the ones for the measurements with `stale=false`.
This makes the configuration easier to reason about and also the code simpler

This commit ports only parts of the latency tests. The other latency test
will be ported in a subsequent commit.

Change-Id: I7cba6d693792ab73b12a53db063487e77626f947