MB-15514: Remove old mapreduce view tests
There's now a new code path and configuration files for the mapreduce views.
The new code shares a lot with the spatial views which makes maintenance for
both easier.

Also internally a lot of the code changed from `query` and `index` to `mapreduce`
which makes the code easier to understand as we now have several ways to query
and index in Couchbase.

Change-Id: Ia84e18bca8dd10a1feeeaf9cd8ef939036b9ba97